refactor: 상품 화면 컴포넌트 파일 분리
