+++ app/src/main/java/nextstep/shoppingcart/ui/screen/component/ProductItem.kt
@@ -1,0 +1,43 @@
+package nextstep.shoppingcart.ui.screen.component
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+import coil3.compose.AsyncImage
+
+@Composable
+fun ProductItem(
+    title: String,
+    price: Int,
+    imageUrl: String,
+    modifier: Modifier = Modifier,
+) {
+    Column(
+        modifier = modifier,
+        horizontalAlignment = Alignment.Start,
+    ) {
+        AsyncImage(
+            model = imageUrl,
+            contentDescription = "Product Image",
+            contentScale = ContentScale.Crop,
+            modifier = Modifier
+                .fillMaxWidth()
+                .aspectRatio(78f / 79f)
+        )
+        Text(
+            text = title,
+            fontSize = 16.sp,
+            fontWeight = FontWeight.W700,
+        )
+        Text(
+            text = "${price}원",
+        )
+    }
+}
