+++ app/src/main/java/nextstep/shoppingcart/ui/screen/component/TopAppBar.kt
@@ -1,0 +1,39 @@
+package nextstep.shoppingcart.ui.screen.component
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ShoppingCart
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import nextstep.shoppingcart.R
+
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun TopAppBar(
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit
+) {
+    CenterAlignedTopAppBar(
+        title = {
+            Text(text = stringResource(id = R.string.appbar_product_title))
+        },
+        actions = {
+            IconButton(
+                onClick = onClick
+            ) {
+                Icon(
+                    imageVector = Icons.Filled.ShoppingCart,
+                    contentDescription = "Cart",
+                    tint = Color.Black
+                )
+            }
+        }
+    )
+}
